+++ components/shared/Form/Context/Context.js
@@ -1,0 +1,6 @@
+import React from 'react'
+
+export const {Consumer, Provider} = React.createContext({
+  value: {},
+  onChange: () => () => null
+})
