+++ components/shared/Form/fields/TextInput.js
@@ -1,6 +1,6 @@
 import TextInput from '@/components/shared/TextInput'
-import {field} from '../Context'
+import {field} from '../Context/Consumer'
 
-export default field(({onChange, ...props}) => (
+export default field()(({onChange, ...props}) => (
   <TextInput {...props} onChangeText={onChange} />
 ))
