+++ components/shared/Form/index.js
@@ -1,5 +1,5 @@
-export {default, field} from './Context'
-export {default as Validator, validate} from '/Validator'
-// Fields
+export {default} from './Context/Provider'
+export {pureField, field} from './Context/Consumer'
+
 export {default as Field} from './Field'
 export {default as TextInput} from './fields/TextInput'
